webapi/tct-transitions-css3-tests/transitions/CSS3Transitions_transition_4s_1s.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        haox.li <haox.li@intel.com>
        Xinx, liu <xinx.liu@intel.com>

-->

<html>
  <head>
    <title>CSS3 Transitions Test: CSS3Transitions_transition_4s_1s</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://www.w3.org/TR/2012/WD-css3-transitions-20120403/" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if the first value that can be parsed as a time is assigned to the transition-duration,the second is transition-delay" />
    <script src="support/support.js"></script>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <style>
        #test {
            height: 100px;
            width: 100px;
            background: blue;
            border: 5px solid red;
        }
    </style>
  </head>
  <body>
    <div id="log"></div>
    <div id="test"></div>
    <button href="#" onclick = "into()">click</button>
    <script>
        function into() {
            var newid = document.getElementById("test");
            newid.style.height = "200px";
            newid.style.width = "200px";
        }
        var div = document.querySelector("#test");
        div.style[headProp("transition")] = "all 4s ease 1s";
        var test = GetCurrentStyle("transitionDuration");
        var test1 = GetCurrentStyle("transitionDelay");
        var t = async_test(document.title, {timeout: 500});
        t.step(function () {
            assert_equals(test, "4s", "The Transitions transition-duration ");
            assert_equals(test1, "1s", "The Transitions transition-delay ");
        });
        t.done();
    </script>
  </body>
</html>
